+++ lib/models/auth/auth.dart
@@ -1,0 +1,3 @@
+library geoff.models.auth;
+
+export 'geoff_user.dart';