+++ lib/models/models.dart
@@ -1,1 +1,5 @@
 library geoff.models;
+
+export 'model_utils.dart';
+export './auth/auth.dart';
+export './exceptions/exceptions.dart';